+++ src/Light/DirectionalLight.js
@@ -33,5 +33,9 @@ WHS.DirectionalLight = class DirectionalLight extends WHS.Light {
 }
 
 WHS.World.prototype.DirectionalLight = function( params ) {
-	return ( new WHS.DirectionalLight(  params ) ).addTo( this );
+    let object = new WHS.DirectionalLight( params );
+
+    object.addTo( this );
+
+    return object;
 }